Gracefully skip unsupported generics constructs.
For example, model classes with return types of `Callable<?>` should
be processed without error. The `<?>` wildcard type data is lost, but
the field should show up in at least basic form.

This behavior will apply to intersection and union types, and other
unsupported generics constructs.

PRs to enhance this behavior are welcome.