Match Jackson's usage of treating Optional<T> as T
When using Jackson's Jdk8Module, fields that have a null value are
deserialized as Optional.empty(), whereas a field that is not provided
will simply be null. This allows REST endpoints to distinguish between
the user not providing a field vs providing a null value without needing
to implement special logic in the bean builder. This is especially
useful for PATCH requests that may want to allow for either setting a
field to null *or* leaving it at its current value by omitting the
field.

Prior to this commit, using this pattern would result in wsdoc
traversing the Optional class itself, which is not useful. This commit
causes Optional<T> to be simply treated as T, matching Jackson's
behavior.

The tests for this functionality are pattered after the tests for the
UUID type, which also has fields that should not be traversed.